Renamed data dir to types. Split types file up into separate files for entity and citation types.
