+++ ui/src/types/entity.d.ts
@@ -1,0 +1,13 @@
+import { BoundingBox } from '../library/types';
+
+export enum ENTITY_TYPE {
+  CITATION = 'citation',
+}
+
+export type Entity = {
+  id: string;
+  type: ENTITY_TYPE;
+  attributes: {
+    boundingBoxes: Array<BoundingBox>;
+  };
+};
